modelo pronto

+++ app/Http/Controllers/UserController.php
@@ -5,13 +5,22 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use App\Models\UserProfile;
 use App\Models\Role;
+use App\Http\Controllers\Gate;
 
 use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
-    public function index(){
-        $users = User::all();
+    public function index(Request $request){
+        $users = User::query();
+        $users->when($request->keyword, function($query, $keyword) {
+            $query->where(function($q) use($keyword){
+                $q->where('name', 'ilike', '%' .$keyword .'%')
+                ->orWhere('email', 'ilike', '%' .$keyword .'%');
+            });
+        });
+
+        $users = $users->orderBy('name', 'desc')->paginate(2);
         return view('users.index', compact('users'));
     }
 
@@ -28,19 +37,25 @@ class UserController extends Controller
 
         User::create($users);
 
-        return back()
+        return redirect(Route('users.index', $user->id))
         //->Route('users.index')
             ->with('status', 'Usuário adicionado com sucesso!');
     }
 
     public function edit(User $user){
+        Gate::authorize('edit', User::class);
         //Vai buscar a relação profile no model User
         $user->load(['profile', 'interests']);
         $roles = Role::all();
         return view('users.edit', compact('user', 'roles'));
+
+        return redirect(Route('users.index', $user->id))
+            //->Route('users.edit', $user->id)
+            ->with('status', 'Usuário editado com sucesso!');
     }
 
     public function update(User $user, Request $request){
+        Gate::authorize('edit', User::class);
         $users =$request->validate([
             'name' => 'required',
             'email' => 'required|email',
@@ -50,12 +65,13 @@ class UserController extends Controller
         $user->fill($users);
         $user->save();
 
-        return back()
+        return redirect(Route('users.index', $user->id))
             //->Route('users.edit', $user->id)
             ->with('status', 'Usuário editado com sucesso!');
     }
 
     public function updateProfile(User $user, Request $request){
+        Gate::authorize('edit', User::class);
         $users =$request->validate([
             'type' => 'required',
             'address' => 'nullable',
@@ -72,12 +88,13 @@ class UserController extends Controller
         //Irá criar um novo registro toda vez
         //$user->profile()->create($request);
 
-        return back()
+        return redirect(Route('users.index', $user->id))
             //->Route('users.edit', $user->id)
             ->with('status', 'Usuário editado com sucesso!');
     }
 
     public function updateInterests(User $user, Request $request){
+        Gate::authorize('edit', User::class);
         $users =$request->validate([
             'interests' => 'nullable|array'
         ]);
@@ -88,22 +105,29 @@ class UserController extends Controller
             $user->interests()->createMany($users['interests']);
         }
 
-        return back()
-            ->with('status', 'Usuário deletado com sucesso!');
+        return redirect(Route('users.index', $user->id))
+            ->with('status', 'Usuário editado com sucesso!');
     }
 
     public function updateRoles(User $user, Request $request){
+        Gate::authorize('edit', User::class);
         $users =$request->validate([
-            'roles' => 'request|array'
+            'roles' => 'required|array'
         ]);
 
-        dd($users->all());
+        // No Model User, pegamos a relação roles e usamos o método attach para relacionar as duas tabelas
+        //Alterando o attach para sync, ele deleta tudo da tabela e registra novamente, para evitar duplicidades
+        $user->roles()->sync($users['roles']);
+
+        return redirect(Route('users.index', $user->id))
+            ->with('status', 'Usuário editado com sucesso!');
     }
 
     public function destroy(User $user){
+        Gate::authorize('destroy', User::class);
         $user->delete();
 
-        return back()
+        return redirect(Route('users.index', $user->id))
             ->with('status', 'Usuário deletado com sucesso!');
     }
